requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-PyMongo==2.3.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jwt==1.2.0
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==1.4.2
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
validate-email==1.3
Werkzeug==2.0.1
